Always try to load zfs module on boot
Often the zfs module is not loaded automatically, which causes qubesd to not start
(because the zfs storage module fails to find the zfs volumes and throws an exception),
which again causes all manner of issues, so even if you manually `modprobe zfs` and restart `qubesd`,
the GUI daemon will still not work correctly.

This commit adds a file containing "zfs" to /etc/modules-load.d/zfs-for-qubes.conf to attempt
unconditional loading of the zfs module.